Split out shared verification logic
